Resolve 64 bit gcc build error at win: src/utils/svalue.cpp:288:38: error: cast from 'qf::core::utils::SValue::Data*' to 'long int' loses precision [-fpermissive]

+++ libqf/libqfcore/src/utils/svalue.cpp
@@ -285,7 +285,7 @@ void SValue::setProperty(int ix, const QVariant &val)
 
 QString SValue::dataHex() const
 {
-	return QString::number((long)d.data(), 16);
+        return QString::number((intptr_t)d.data(), 16);
 }
 
 QVariant SValue::variantCast() const
